[refactor] Moved deno logic into deno service
